Handle endpoints with the same name
AWS has introduced duplication in endpoint names in the metadata such as
iam/info and identity-credentials/ec2/info. Previously the first endpoint
found would get added to the CLI options and the second one would get dropped.
The new code will use the path to differentiate the query options and collect
all the data when no query option is specified.