Eliminate ambigutity
When an endpoint exists expand the option name of the existing end point and
the endpoint found with the same name. Remove the existing entry. This avoids
providing the same information with 2 options and eliminates order dependency
on the metadata information as the first entry found would have occupied the
conflicting name which may have triggered flip-flop.