Switch the package build to a specified Python interpreter
In openSUSE Factory we are working to eliminate all python3- packages. For
SLE we can switch over to a Python 3.11 build starting with 15 SP4. If we need
updates in code bases prior to 15 SP4 we will address those as package patches.
For distributions >16 we use the define primary Python.

While the package also provides a library we do not want to support builds
with multiple interpreters. Code that uses this package as library will have
to build with the specific interpreter for the distribution. The reason for
this is that we do not want to change our image build setup every time the
distribution changes the Python interpreter. A universal build setup
produces python$PYVERSION-$NAME binary packages and we do not want to
chase this in our image build setups.